refactor: move/rename components for admin, auth, library, and student
