+++ src/features/admin/pages/DepartmentRequestPage/DepartmentRequestPage.tsx
@@ -1,17 +1,17 @@
 import { Footer } from "@/components/layout/Footer/Footer";
 import { Header } from "@/components/layout/Header/Header";
-import { RequestTable } from "../components/RequestTable/RequestTable";
-import style from "./RequestPage.module.css";
+import style from "./DepartmentRequestPage.module.css";
 
-export const RequestPage = () => {
+export const DepartmentRequestPage = () => {
   return (
     <div className={style.page}>
       <Header />
       <main className={style.main}>
         <div className={style.mainContainer}>
-          <section className={style.tableSection}>
-            <RequestTable />
-          </section>
+          <h1 className={style.titleHeader}>Manage Document Requests (Department Admin) </h1>
+          <div className={style.loadingContainer}>
+            <section className={style.tableSection} />
+          </div>
         </div>
       </main>
       <Footer />
