+++ src/features/admin/pages/DepartmentResearchPage/DepartmentResearchPage.module.css
