+++ src/features/admin/pages/SuperRequestPage/SuperRequestPage.tsx
@@ -1,9 +1,9 @@
 import { Footer } from "@/components/layout/Footer/Footer";
 import { Header } from "@/components/layout/Header/Header";
 import { RequestsTable } from "@/features/admin/components/RequestTable/RequestTable";
-import style from "./RequestPage.module.css";
+import style from "./SuperRequestPage.module.css";
 
-export const RequestPage = () => {
+export const SuperRequestPage = () => {
   return (
     <div className={style.page}>
       <Header />
